refactor(source-edit): move leftover root edit glue into the crate
Root edits/ and move_symbol/ orchestration now live in
tracedecay-source-edit and call admitted SourceEditGraphReadV1.
