+++ crates/tracedecay-source-edit/Cargo.toml
@@ -10,16 +10,20 @@ repository = "https://github.com/ScriptedAlchemy/tracedecay"
 [dependencies]
 cap-fs-ext = "4.0.2"
 cap-std = "4.0.2"
+gix = { version = "=0.86.0", default-features = false, features = ["revision", "blob-diff", "parallel", "sha1", "sha256", "status"] }
 hotpath.workspace = true
 same-file = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = "0.11"
+tempfile = "3"
 tokio = { version = "1", features = ["rt"] }
+tree-sitter = "0.26"
 tracing = "0.1"
 tracedecay-contracts = { path = "../tracedecay-contracts", version = "0.1.0" }
-# Grammar-free entry only: the admitted graph handle wraps the interactive
-# reader, so no tree-sitter bundle is linked into this crate.
+# The admitted graph handle wraps the interactive reader. Move-symbol import
+# analysis needs the Rust grammar so `try_language("rust")` is registered.
+tracedecay-code-extraction = { path = "../tracedecay-code-extraction", version = "0.1.0", default-features = false, features = ["lite"] }
 tracedecay-code-index = { path = "../tracedecay-code-index", version = "0.1.0", default-features = false }
 tracedecay-domain = { path = "../tracedecay-domain", version = "0.1.0" }
 tracedecay-graph-db = { path = "../tracedecay-graph-db", version = "0.1.0" }
