+++ crates/tracedecay-source-edit/src/edits/reconcile_tests.rs
@@ -6,7 +6,7 @@ use tracedecay_contracts::{
 use tracedecay_domain::UtcMicros;
 
 use super::test_support::effect_unknown_fixture;
-use tracedecay_source_edit::{
+use crate::{
     SourceEditEffectControlV1, execute_source_edit,
     reconcile_source_edit_effect_unknown_with_control,
 };
