+++ crates/tracedecay-source-edit/src/edits/rename/graph_evidence.rs
@@ -1,6 +1,7 @@
 use std::collections::{BTreeMap, BTreeSet};
 use std::sync::Arc;
 
+use crate::SourceEditGraphReadV1;
 use tracedecay_code_index::graph_projection::{
     CodeGraphProjectionError, CodeGraphSemanticEdgeV1, CodeGraphSymbolSummaryV1,
 };
@@ -12,7 +13,6 @@ use tracedecay_domain::{
     SourceSpan, SymbolOccurrenceId, canonical_sha256,
 };
 use tracedecay_graph_query::{map_code_graph_read_runtime_error, map_projection_error};
-use tracedecay_source_edit::SourceEditGraphReadV1;
 
 use tracedecay_domain::errors::{Result, TraceDecayError};
