+++ crates/tracedecay-source-edit/src/move_symbol/hints.rs
