+++ crates/tracedecay-source-edit/src/move_symbol/rust_paths.rs
