--- crates/tracedecay/src/tracedecay/edits/mod.rs
@@ -1,85 +1,0 @@
-//! Anchored source-editing primitives (str-replace, insert, symbol
-//! replacement, ast-grep rewrites) plus the single-file re-index they
-//! trigger.
-//!
-//! Direct graph mutations are crate-internal adapters; external callers must
-//! use the canonical source-edit transaction.
-//!
-//! ```compile_fail
-//! async fn direct_str_replace_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph.str_replace("src/lib.rs", "old", "new", true).await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_multi_str_replace_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph
-//!         .multi_str_replace("src/lib.rs", &[("old", "new")], true)
-//!         .await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_insert_at_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph
-//!         .insert_at("src/lib.rs", "anchor", "content", true, true)
-//!         .await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_replace_symbol_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph.replace_symbol("symbol", "fn symbol() {}", true).await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_insert_at_symbol_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph
-//!         .insert_at_symbol("symbol", "content", "before", true)
-//!         .await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_ast_grep_rewrite_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph
-//!         .ast_grep_rewrite("src/lib.rs", "$A", "$A", true)
-//!         .await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_move_symbol_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
-//!     let _ = graph
-//!         .move_symbol("symbol", "src/dest.rs", true, false)
-//!         .await;
-//! }
-//! ```
-//! ```compile_fail
-//! async fn direct_rename_symbol_is_not_public(
-//!     graph: &tracedecay::tracedecay::TraceDecay,
-//!     binding: &tracedecay_contracts::RenameSymbolBindingV1,
-//! ) {
-//!     let _ = graph.rename_symbol(binding, "new_name", true).await;
-//! }
-//! ```
-
-mod ast_grep;
-mod preview;
-mod primitives;
-mod rename;
-mod symbols;
-
-#[cfg(test)]
-mod execute_tests;
-#[cfg(test)]
-mod reconcile_tests;
-#[cfg(test)]
-mod test_support;
-
-// `move_symbol.rs` (a sibling of this module) imports these edit helpers via
-// `use super::edits::{...}`; the re-exports resolve each name at
-// `crate::tracedecay::edits`.
-pub(in crate::tracedecay) use preview::{
-    LeadingKind, MAX_PREVIEW_DIFF_LINES, PREVIEW_DIFF_CONTEXT, bounded_region_diff,
-    classify_leading_line, edit_success_message,
-};
-pub(in crate::tracedecay) use primitives::splice_lines;
-pub(in crate::tracedecay) use symbols::{
-    EditSymbolV1, edit_symbol_from_summary, resolve_symbol_for_edit,
-};
