+++ crates/tracedecay/src/tracedecay/source_edit_runtime.rs
@@ -1,14 +1,65 @@
+//! Direct graph mutations are crate-internal adapters; external callers must
+//! use the canonical source-edit transaction.
+//!
+//! ```compile_fail
+//! async fn direct_str_replace_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph.str_replace("src/lib.rs", "old", "new", true).await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_multi_str_replace_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph
+//!         .multi_str_replace("src/lib.rs", &[("old", "new")], true)
+//!         .await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_insert_at_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph
+//!         .insert_at("src/lib.rs", "anchor", "content", true, true)
+//!         .await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_replace_symbol_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph.replace_symbol("symbol", "fn symbol() {}", true).await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_insert_at_symbol_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph
+//!         .insert_at_symbol("symbol", "content", "before", true)
+//!         .await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_ast_grep_rewrite_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph
+//!         .ast_grep_rewrite("src/lib.rs", "$A", "$A", true)
+//!         .await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_move_symbol_is_not_public(graph: &tracedecay::tracedecay::TraceDecay) {
+//!     let _ = graph
+//!         .move_symbol("symbol", "src/dest.rs", true, false)
+//!         .await;
+//! }
+//! ```
+//! ```compile_fail
+//! async fn direct_rename_symbol_is_not_public(
+//!     graph: &tracedecay::tracedecay::TraceDecay,
+//!     binding: &tracedecay_contracts::RenameSymbolBindingV1,
+//! ) {
+//!     let _ = graph.rename_symbol(binding, "new_name", true).await;
+//! }
+//! ```
+
 use std::path::Path;
 
-use tracedecay_contracts::source_edit::{
-    AstGrepResult, EditResult, InsertResult, MoveResult, MultiEditResult, RenameResult,
-    RenameSymbolBindingV1,
-};
 use tracedecay_domain::errors::TraceDecayError;
 use tracedecay_runtime_core::storage::StoreLayout;
-use tracedecay_source_edit::{
-    EditDiagnosticRecord, SourceEditFuture, SourceEditGraphReadV1, SourceEditRuntimePort,
-};
+use tracedecay_source_edit::{EditDiagnosticRecord, SourceEditFuture, SourceEditRuntimePort};
 
 use super::TraceDecay;
 
@@ -32,111 +83,5 @@ impl SourceEditRuntimePort for TraceDecay {
                 "source-edit verification requires the daemon-owned LSP diagnostics authority",
             ))
         })
-    }
-
-    fn replace_symbol<'a>(
-        &'a self,
-        graph: SourceEditGraphReadV1,
-        symbol: &'a str,
-        new_source: &'a str,
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, EditResult> {
-        Box::pin(TraceDecay::replace_symbol(
-            self, graph, symbol, new_source, dry_run,
-        ))
-    }
-
-    fn str_replace<'a>(
-        &'a self,
-        path: &'a str,
-        old_str: &'a str,
-        new_str: &'a str,
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, EditResult> {
-        Box::pin(TraceDecay::str_replace(
-            self, path, old_str, new_str, dry_run,
-        ))
-    }
-
-    fn multi_str_replace<'a>(
-        &'a self,
-        path: &'a str,
-        replacements: &'a [(&'a str, &'a str)],
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, MultiEditResult> {
-        Box::pin(TraceDecay::multi_str_replace(
-            self,
-            path,
-            replacements,
-            dry_run,
-        ))
-    }
-
-    fn insert_at<'a>(
-        &'a self,
-        path: &'a str,
-        anchor: &'a str,
-        content: &'a str,
-        before: bool,
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, InsertResult> {
-        Box::pin(TraceDecay::insert_at(
-            self, path, anchor, content, before, dry_run,
-        ))
-    }
-
-    fn insert_at_symbol<'a>(
-        &'a self,
-        graph: SourceEditGraphReadV1,
-        symbol: &'a str,
-        content: &'a str,
-        position: &'a str,
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, InsertResult> {
-        Box::pin(TraceDecay::insert_at_symbol(
-            self, graph, symbol, content, position, dry_run,
-        ))
-    }
-
-    fn ast_grep_rewrite<'a>(
-        &'a self,
-        path: &'a str,
-        pattern: &'a str,
-        rewrite: &'a str,
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, AstGrepResult> {
-        Box::pin(TraceDecay::ast_grep_rewrite(
-            self, path, pattern, rewrite, dry_run,
-        ))
-    }
-
-    fn move_symbol<'a>(
-        &'a self,
-        graph: SourceEditGraphReadV1,
-        symbol: &'a str,
-        dest_file: &'a str,
-        dry_run: bool,
-        update_references: bool,
-    ) -> SourceEditFuture<'a, MoveResult> {
-        Box::pin(TraceDecay::move_symbol(
-            self,
-            graph,
-            symbol,
-            dest_file,
-            dry_run,
-            update_references,
-        ))
-    }
-
-    fn rename_symbol<'a>(
-        &'a self,
-        graph: SourceEditGraphReadV1,
-        binding: &'a RenameSymbolBindingV1,
-        new_name: &'a str,
-        dry_run: bool,
-    ) -> SourceEditFuture<'a, RenameResult> {
-        Box::pin(TraceDecay::rename_symbol(
-            self, graph, binding, new_name, dry_run,
-        ))
     }
 }
